+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add property `title` to Pie-Chart
 - Add properties `xlabel` and `ylabel` to Line-Chart
 - Add properties `xlabel` and `ylabel` to Histogram
+- Add properties `xlabel` and `ylabel` to Scatter-Chart
 
 ## [0.2.2] - 2019-10-06